clean up hardhat config file and move tasks to folder
